@@ -2,7 +2,7 @@ package ru.practicum.main.category.model;
 
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Getter
 @Setter
