@@ -12,8 +12,8 @@ import ru.practicum.main.event.dto.EventFullDto;
 import ru.practicum.main.event.service.EventAdminService;
 import ru.practicum.main.event.status.State;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.time.LocalDateTime;
 import java.util.List;
 
