@@ -11,8 +11,8 @@ import ru.practicum.main.event.dto.*;
 import ru.practicum.main.event.service.EventPrivateService;
 import ru.practicum.main.request.dto.RequestDto;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.util.List;
 
 @RestController
