@@ -2,7 +2,7 @@ package ru.practicum.main.stat.service;
 
 import ru.practicum.main.event.model.Event;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.Collection;
 import java.util.Map;
 
